Create main (Доход с учетом фиксированной ставки налога).cpp
This repository contains a simple C++ program that demonstrates the use of static member variables and static member functions to calculate income tax. The program includes a TaxCalculator class with static methods to calculate and retrieve the income tax based on predefined income and tax rate. This example is designed to help beginners understand the use of static members and functions in C++.